refactor(gateway): extract silence-poke session-event dispatch to a module (#3521 ratchet)
After rebasing onto main (post-#3523), gateway.ts hit 24920 lines, 3 over the
anti-inflation ratchet ceiling (24867 + 50 grace). Rather than bump the ceiling,
extract the inline silence-poke session-event dispatch block (thinking /
tool_use / tool_result branches + the #3519 background-shell liveness call) out
of onSessionEvent into a new sibling module,
telegram-plugin/gateway/silence-poke-session-event.ts, exposing
applySilencePokeSessionEvent(). Consolidates #3521's applyBackgroundShellLiveness
call with the sibling silence-poke wiring it participates in, following the
existing modularization pattern.

Pure extract/move — behavior byte-for-byte identical. gateway.ts 24920 -> 24863
(4 under ceiling). Removed now-unused toolLabel / applyBackgroundShellLiveness
imports from gateway.ts.

Lint (incl. ratchet) clean; scoped silence-poke / liveness / session-tail tests
174/174 pass.
