+++ telegram-plugin/gateway/gateway.ts
@@ -276,7 +276,6 @@ import { createSessionModelSource } from './session-model-source.js'
 import { runSilentTurnHeartbeatTick } from '../feed-heartbeat-climb.js'
 import { REPLY_TOOLS } from '../narrative-dedup.js'
 import { NarrativeFlushController, PENDING_NARRATIVE_FLUSH_MS } from '../narrative-flush.js'
-import { toolLabel } from '../tool-labels.js'
 import { createTypingWrapper } from '../typing-wrap.js'
 import { createTurnTypingLoop } from './turn-typing-loop.js'
 import {
@@ -752,7 +751,7 @@ import { createDeliveryConfirmWiring } from './delivery-confirm-wiring.js'
 import { createObligationWiring } from './obligation-wiring.js'
 // #2996 P8 PR-C3 — the extracted silence-poke wiring.
 import { buildSilencePokeOptions } from './liveness-wiring.js'
-import { applyBackgroundShellLiveness } from './background-shell-liveness.js'
+import { applySilencePokeSessionEvent } from './silence-poke-session-event.js'
 import type { ChatKey as _ChatKey } from './inbound-delivery-machine.js'
 import { dispatchEffects } from './inbound-delivery-machine-dispatch.js'
 import { maybeFireWarmup } from './prefix-warmup.js'
@@ -10873,63 +10872,7 @@ if (isGatewayMain) ipcServer = createIpcServer({
     // (thinking vs working, plus the longest-running in-flight tool).
     if (currentTurn != null) {
       const key = statusKey(currentTurn.sessionChatId, currentTurn.sessionThreadId)
-      if (ev.kind === 'thinking') {
-        silencePoke.noteThinking(key, Date.now())
-      } else if (ev.kind === 'tool_use') {
-        // #1292: track in-flight tool calls so the 300s framework
-        // fallback message can name the actual observable (e.g.
-        // "running Grep \"foo\" for 4m") instead of the dishonest
-        // generic "still working… no update in 5 min" when the agent
-        // is clearly busy on tool calls. Telegram-surface tools are
-        // excluded — their job IS the outbound message, the silence
-        // clock resets via noteOutbound when they fire. Sub-agent
-        // tool_use events (kind='sub_agent_tool_use') intentionally
-        // NOT tracked: the parent's Task tool_use is already on the
-        // map and represents the user-observable wait.
-        if (
-          ev.toolUseId != null
-          && ev.toolUseId.length > 0
-          && !isTelegramSurfaceTool(ev.toolName)
-        ) {
-          const label = toolLabel(
-            ev.toolName,
-            ev.input,
-            /*preamble*/ undefined,
-            ev.precomputedLabel,
-          )
-          silencePoke.noteToolStart(
-            key,
-            ev.toolUseId,
-            ev.toolName,
-            label.length > 0 ? label : null,
-            Date.now(),
-          )
-          // #1445 cross-turn pending-async ambient. Mark the chat as
-          // having dispatched background work this turn so a turn_end
-          // that follows activates the edit-in-place ambient line.
-          // Covers `Agent` / `Task` (the harness-managed async path
-          // — handback channel turn clears it) and `Bash` with
-          // run_in_background:true (model is expected to poll
-          // BashOutput; the ambient ticks until next inbound or the
-          // 30-min budget cap).
-          const evInput = ev.input as { run_in_background?: boolean } | undefined
-          if (
-            ev.toolName === 'Agent'
-            || ev.toolName === 'Task'
-            || (ev.toolName === 'Bash' && evInput?.run_in_background === true)
-          ) {
-            pendingProgress.noteAsyncDispatch(key)
-          }
-        }
-      } else if (ev.kind === 'tool_result') {
-        // #1292: drain the in-flight entry. Idempotent on unknown ids
-        // (covers Telegram-surface tools we skipped at start time).
-        if (ev.toolUseId != null && ev.toolUseId.length > 0) {
-          silencePoke.noteToolEnd(key, ev.toolUseId, Date.now())
-        }
-      }
-      // #3519 sharpen: feed background-shell liveness (ALIVE/DEAD) to silence-poke — see background-shell-liveness.ts.
-      applyBackgroundShellLiveness(silencePoke, key, ev)
+      applySilencePokeSessionEvent(silencePoke, pendingProgress, key, ev)
     }
   },
